requirements.txt
aioredis==1.3.1
asgiref==3.2.3
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
certifi==2019.6.16
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
constantly==15.1.0
crayons==0.3.0
cryptography==2.8
daphne==2.4.1
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.11.0
docker==4.1.0
gspread==3.1.0
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.13.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
msgpack==0.6.2
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.12
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
rsa==4.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.3.1
urllib3==1.25.3
webdriver-manager==2.3.0
websocket-client==0.57.0
wrapt==1.11.1
zope.interface==4.7.1